64-bit MemoryOps
MemoryOps has some helper functions to copy or fill memory blocks is a few
special ways. Before this patch the size of such a blocks was limited to 32
bit, now they support 64-bit sizes.

Likely (hopefully) we'll never actually have to copy/fill such huge blocks,
but there's again no good reason to limit the possibility.

Some of these routines use x86 inline assembly. With gcc inline assembly syntax
it's possible to write one routine that compiles fine to either 32-bit or
64-bit x86 code. With visual studio this is not possible (at least I don't know
how). The code did contain a visual studio inline asm routine that should
be adapted to work with 64-bit types (on 64-bit systems). Though that routine
was already commented out. So instead of (almost) duplicating it for 64-bit
platforms I just removed it.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13082 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/utils/MemoryOps.hh
@@ -4,6 +4,7 @@
 #define MEMORYOPS_HH
 
 #include "openmsx.hh"
+#include <cstring> // for size_t
 
 namespace openmsx {
 namespace MemoryOps {
@@ -12,10 +13,10 @@ namespace MemoryOps {
 	static const bool STREAMING    = true;
 
 	template <typename Pixel, bool STREAMING> struct MemSet {
-		void operator()(Pixel* out, unsigned num, Pixel val) const;
+		void operator()(Pixel* out, size_t num, Pixel val) const;
 	};
 	template <typename Pixel, bool STREAMING> struct MemSet2 {
-		void operator()(Pixel* out, unsigned num,
+		void operator()(Pixel* out, size_t num,
 		                Pixel val0, Pixel val1) const;
 	};
 
@@ -28,10 +29,10 @@ namespace MemoryOps {
 	  *            different from the normal memcpy function, there this
 	  *            parameter indicates the number of bytes.
 	  */
-	void stream_memcpy(unsigned* dst, const unsigned* src, unsigned num);
-	void stream_memcpy(word* dst, const word* src, unsigned num);
+	void stream_memcpy(unsigned* dst, const unsigned* src, size_t num);
+	void stream_memcpy(word* dst, const word* src, size_t num);
 
-	void* mallocAligned(unsigned alignment, unsigned size);
+	void* mallocAligned(size_t alignment, size_t size);
 	void freeAligned(void* ptr);
 
 } // namespace MemoryOps
